+++ CHANGELOG.md
@@ -4,6 +4,15 @@
 This file contains the list of changes made to the JLS project.
 
 
+## 0.10.0
+
+2024 Aug 21
+
+* Improved Python binding to simplify finding sources and signals.
+  * Added pyjls.source_lookup.
+  * Improved pyjls.signal_lookup to support '<source_name>.<signal_name>' format.
+
+
 ## 0.9.7
 
 2024 Jul 25